@@ -2,7 +2,7 @@
 
 # https://mkdocstrings.github.io/recipes/
 """
-from __future__ import annotations
+
 
 import logging
 from pathlib import Path
