@@ -1,4 +1,4 @@
-from __future__ import annotations
+
 
 from functools import partial
 from typing import Callable
