@@ -1,17 +1,15 @@
 # type: ignore
-from __future__ import annotations
+
 
 from copy import deepcopy
-
-import numpy as np
-import pandas as pd
 import torch
 
 from neps.utils.common import instance_from_map
-from ..bayesian_optimization.models import SurrogateModelMapping
-from ..multi_fidelity.utils import normalize_vectorize_config
-from ..multi_fidelity_prior.utils import calc_total_resources_spent, update_fidelity
-from ..utils import map_real_hyperparameters_from_tabular_ids
+from neps.optimizers.bayesian_optimization.models import SurrogateModelMapping
+from neps.optimizers.multi_fidelity.utils import (
+    get_tokenized_data, get_training_data_for_freeze_thaw
+)
+from neps.optimizers.multi_fidelity_prior.utils import calc_total_resources_spent, update_fidelity
 
 
 class MFBOBase:
@@ -142,7 +140,7 @@ class MFBOBase:
     def sample_new_config(
         self,
         rung: int = None,
-        **kwargs,
+        **kwargs,  # pylint: disable=unused-argument
     ):
         """Samples configuration from policies or random."""
         if self.model_based and not self.is_init_phase():
@@ -187,8 +185,9 @@ class FreezeThawModel:
     def __init__(
         self,
         pipeline_space,
-        surrogate_model: str = "deep_gp",
+        surrogate_model: str = "ftpfn",
         surrogate_model_args: dict = None,
+        step_size: int = 1,
     ):
         self.observed_configs = None
         self.pipeline_space = pipeline_space
@@ -196,103 +195,40 @@ class FreezeThawModel:
         self.surrogate_model_args = (
             surrogate_model_args if surrogate_model_args is not None else {}
         )
-        if self.surrogate_model_name in ["deep_gp", "pfn"]:
-            self.surrogate_model_args.update({"pipeline_space": pipeline_space})
-
-        # instantiate the surrogate model
         self.surrogate_model = instance_from_map(
             SurrogateModelMapping,
             self.surrogate_model_name,
             name="surrogate model",
             kwargs=self.surrogate_model_args,
         )
+        self.step_size = step_size
 
     def _fantasize_pending(self, train_x, train_y, pending_x):
-        # Select configs that are neither pending nor resulted in error
-        completed_configs = self.observed_configs.completed_runs.copy(deep=True)
-        # IMPORTANT: preprocess observations to get appropriate training data
-        train_x, train_lcs, train_y = self.observed_configs.get_training_data_4DyHPO(
-            completed_configs, self.pipeline_space
-        )
-        pending_condition = self.observed_configs.pending_condition
-        if pending_condition.any():
-            pending_configs = self.observed_configs.df.loc[pending_condition]
-            pending_x, pending_lcs, _ = self.observed_configs.get_training_data_4DyHPO(
-                pending_configs
-            )
-            self._fit(train_x, train_y, train_lcs)
-            _y, _ = self._predict(pending_x, pending_lcs)
-            _y = _y.tolist()
-
-            train_x.extend(pending_x)
-            train_y.extend(_y)
-            train_lcs.extend(pending_lcs)
-
-        return train_x, train_y, train_lcs
+        raise NotImplementedError("Fantasization not implemented yet!")
 
     def _fit(self, train_x, train_y, train_lcs):
-        if self.surrogate_model_name in ["gp", "gp_hierarchy"]:
-            self.surrogate_model.fit(train_x, train_y)
-        elif self.surrogate_model_name in ["deep_gp", "pfn"]:
-            self.surrogate_model.fit(train_x, train_y, train_lcs)
-        else:
-            # check neps/optimizers/bayesian_optimization/models/__init__.py for options
-            raise ValueError(
-                f"Surrogate model {self.surrogate_model_name} not supported!"
-            )
+        raise NotImplementedError("Predict not implemented yet!")
 
-    def _predict(self, test_x, test_lcs):
-        if self.surrogate_model_name in ["gp", "gp_hierarchy"]:
-            return self.surrogate_model.predict(test_x)
-        elif self.surrogate_model_name in ["deep_gp", "pfn"]:
-            return self.surrogate_model.predict(test_x, test_lcs)
-        else:
-            # check neps/optimizers/bayesian_optimization/models/__init__.py for options
-            raise ValueError(
-                f"Surrogate model {self.surrogate_model_name} not supported!"
-            )
+    def _predict(self, test_x) -> torch.Tensor:
+        raise NotImplementedError("Predict not implemented yet!")
 
     def set_state(
         self,
         pipeline_space,
         surrogate_model_args,
-        **kwargs,
+        **kwargs,  # pylint: disable=unused-argument
     ):
         self.pipeline_space = pipeline_space
         self.surrogate_model_args = (
             surrogate_model_args if surrogate_model_args is not None else {}
         )
-        # only to handle tabular spaces
-        if self.pipeline_space.has_tabular:
-            if self.surrogate_model_name in ["deep_gp", "pfn"]:
-                self.surrogate_model_args.update(
-                    {"pipeline_space": self.pipeline_space.raw_tabular_space}
-                )
-            # instantiate the surrogate model, again, with the new pipeline space
-            self.surrogate_model = instance_from_map(
-                SurrogateModelMapping,
-                self.surrogate_model_name,
-                name="surrogate model",
-                kwargs=self.surrogate_model_args,
-            )
-
-    def update_model(self, train_x=None, train_y=None, pending_x=None, decay_t=None):
-        if train_x is None:
-            train_x = []
-        if train_y is None:
-            train_y = []
-        if pending_x is None:
-            pending_x = []
-
-        if decay_t is None:
-            decay_t = len(train_x)
-        train_x, train_y, train_lcs = self._fantasize_pending(
-            train_x, train_y, pending_x
+        self.surrogate_model = instance_from_map(
+            SurrogateModelMapping,
+            self.surrogate_model_name,
+            name="surrogate model",
+            kwargs=self.surrogate_model_args,
         )
-        self._fit(train_x, train_y, train_lcs)
-
-        return self.surrogate_model, decay_t
-
+    
 
 class PFNSurrogate(FreezeThawModel):
     """Special class to deal with PFN surrogate model and freeze-thaw acquisition."""
@@ -302,51 +238,70 @@ class PFNSurrogate(FreezeThawModel):
         self.train_x = None
         self.train_y = None
 
-    def _fit(self, *args):
-        assert self.surrogate_model_name == "pfn"
-        self.preprocess_training_set()
-        self.surrogate_model.fit(self.train_x, self.train_y)
-
-    def preprocess_training_set(self):
-        _configs = self.observed_configs.df.config.values.copy()
-
-        # onlf if tabular space is present
-        if self.pipeline_space.has_tabular:
-            # placeholder index, will be driooed
-            _idxs = np.arange(len(_configs))
-            # mapping the (id, epoch) space of tabular configs to the actual HPs
-            _configs = map_real_hyperparameters_from_tabular_ids(
-                pd.Series(_configs, index=_idxs), self.pipeline_space
-            ).values
-
-        device = self.surrogate_model.device
-        # TODO: fix or make consistent with `tokenize``
-        configs, idxs, performances = self.observed_configs.get_tokenized_data(
-            self.observed_configs.df.copy().assign(config=_configs)
+    def update_model(self):
+        # tokenize the observations
+        idxs, steps, configs, performance = get_training_data_for_freeze_thaw(
+            self.observed_configs.df.loc[self.observed_configs.completed_runs_index],
+            self.observed_configs.config_col,
+            self.observed_configs.perf_col,
+            self.pipeline_space,
+            step_size=self.step_size,
+            maximize=True  # inverts performance since NePS minimizes
         )
-        # TODO: account for fantasization
-        self.train_x = torch.Tensor(np.hstack([idxs, configs])).to(device)
-        self.train_y = torch.Tensor(performances).to(device)
+        df_idxs = torch.Tensor(idxs)
+        df_x = torch.Tensor(get_tokenized_data(configs))
+        df_steps = torch.Tensor(steps)
+        train_x = torch.hstack([
+            df_idxs.reshape(df_steps.shape[0], 1),
+            df_steps.reshape(df_steps.shape[0], 1),
+            df_x
+        ])
+        train_y = torch.Tensor(performance)
+        
+        # fit the model, on only completed runs
+        self._fit(train_x, train_y)
 
-    def preprocess_test_set(self, test_x):
-        _len = len(self.observed_configs.all_configs_list())
-        device = self.surrogate_model.device
+        # fantasize pending evaluations
+        if self.observed_configs.pending_condition.any():
+            # tokenize the pending observations
+            _idxs, _steps, _configs, _ = get_training_data_for_freeze_thaw(
+                self.observed_configs.df.loc[self.observed_configs.pending_runs_index],
+                self.observed_configs.config_col,
+                self.observed_configs.perf_col,
+                self.pipeline_space,
+                step_size=self.step_size,
+                maximize=True  # inverts performance since NePS minimizes
+            )
+            _df_x = torch.Tensor(get_tokenized_data(_configs))
+            _df_idxs = torch.Tensor(_idxs)
+            _df_steps = torch.Tensor(_steps)
+            _test_x = torch.hstack([
+                _df_idxs.reshape(_df_idxs.shape[0], 1),
+                _df_steps.reshape(_df_steps.shape[0], 1),
+                _df_x
+            ])
+            _performances = self._predict(_test_x)  # returns maximizing metric
+            # update the training data
+            train_x = torch.vstack([train_x, _test_x])
+            train_y = torch.hstack([train_y, _performances])
+            # refit the model, on completed runs + fantasized pending runs
+            self._fit(train_x, train_y)
 
-        new_idxs = np.arange(_len, len(test_x))
-        base_fidelity = np.array([1] * len(new_idxs))
-        new_token_ids = np.hstack(
-            (new_idxs.T.reshape(-1, 1), base_fidelity.T.reshape(-1, 1))
-        )
-        # the following operation takes each element in the array and stacks it vertically
-        # in this case, should convert a (n,) array to (n, 2) by flattening the elements
-        existing_token_ids = np.vstack(self.observed_configs.token_ids).astype(int)
-        token_ids = np.vstack((existing_token_ids, new_token_ids))
+    def _fit(self, train_x: torch.Tensor, train_y: torch.Tensor):  # pylint: disable=unused-argument
+        # no training required,, only preprocessing the training data as context during inference
+        assert self.surrogate_model is not None, "Surrogate model not set!"
+        self.surrogate_model.train_x = train_x
+        self.surrogate_model.train_y = train_y
 
-        configs = np.array([normalize_vectorize_config(c) for c in test_x])
-        test_x = torch.Tensor(np.hstack([token_ids, configs])).to(device)
-        return test_x
-
-    def _predict(self, test_x, test_lcs):
-        assert self.surrogate_model_name == "pfn"
-        test_x = self.preprocess_test_set(test_x)
-        return self.surrogate_model.predict(self.train_x, self.train_y, test_x)
+    def _predict(self, test_x: torch.Tensor) -> torch.Tensor:
+        assert self.surrogate_model.train_x is not None and self.surrogate_model.train_y is not None, "Model not trained yet!"
+        if self.surrogate_model_name == "ftpfn":
+            mean = self.surrogate_model.get_mean_performance(test_x)
+            if mean.is_cuda:
+                mean = mean.cpu()
+            return mean
+        else:
+            # check neps/optimizers/bayesian_optimization/models/__init__.py for options
+            raise ValueError(
+                f"Surrogate model {self.surrogate_model_name} not supported!"
+            )
