@@ -1,8 +1,5 @@
-from __future__ import annotations
-
 import logging
-from typing import Any
-from typing_extensions import override
+from typing import Any, override
 
 from neps.state.optimizer import BudgetInfo, OptimizationState
 from neps.utils.types import ConfigResult, RawConfig
