@@ -6,11 +6,12 @@
 from __future__ import annotations
 
 import logging
+from collections.abc import Callable, Iterable, Iterator
 from contextlib import contextmanager
 from copy import deepcopy
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Callable, ClassVar, Generic, Iterable, Iterator, TypeVar
-from typing_extensions import Protocol, Self
+from typing import TYPE_CHECKING, ClassVar, Generic, Protocol, TypeVar
+from typing_extensions import Self
 
 from neps.exceptions import (
     LockFailedError,
