@@ -3,20 +3,19 @@
 from __future__ import annotations
 
 import logging
+from collections.abc import Callable, Mapping
 from dataclasses import asdict, dataclass
 from enum import Enum
-from typing import TYPE_CHECKING, Any, Callable, ClassVar, Literal, Mapping
+from typing import TYPE_CHECKING, Any, ClassVar, Literal
 from typing_extensions import Self
 
 import numpy as np
 
 from neps.exceptions import NePSError
-from neps.utils.types import ConfigResult
+from neps.utils.types import ERROR, ConfigResult, RawConfig
 
 if TYPE_CHECKING:
     from neps.search_spaces import SearchSpace
-    from neps.utils.types import ERROR, RawConfig
-
 
 logger = logging.getLogger(__name__)
 
